requirements.txt
openenv-core @ git+https://github.com/meta-pytorch/OpenEnv.git
gradio
fastapi
uvicorn[standard]
pydantic
openai
